componentize bottomTabNavigator
