+++ src/screens/AccountScreen.tsx
@@ -4,6 +4,7 @@ import { styled } from 'nativewind';
 import { View, Text } from 'react-native';
 
 import Button from '../components/Button';
+import BottomTabNavigator from '../components/BottomTabNavigator';
 
 const StyledView = styled(View);
 const StyledButton = styled(Button);
@@ -24,26 +25,10 @@ const AccountScreen = ({ navigation }) => {
           onPress={() => navigation.navigate('Login')}
         />
       </StyledView>
-      <StyledView className="flex-row pb-10 items-center justify-around h-1/5 w-screen bg-[#345073]">
-        <StyledButton
-          className="flex bg-[#ea5a4e] text-white rounded-3xl w-[120] h-[40]"
-          title="My Cases"
-          titleColor="white"
-          onPress={() => navigation.navigate('CaseContactList')}
-        />
-        <StyledButton
-          className="flex bg-[#ea5a4e] text-white rounded-3xl w-[90] h-[40]"
-          title="Create"
-          titleColor="white"
-          onPress={() => navigation.navigate('CaseContactList')}
-        />
-        <StyledButton
-          className="flex bg-[#ea5a4e] text-white rounded-3xl w-[120] h-[40]"
-          title="Account"
-          titleColor="white"
-          onPress={() => navigation.navigate('AccountScreen')}
-        />
-      </StyledView>
+      <BottomTabNavigator
+        className="flex-row pb-10 items-center justify-around h-1/5 w-screen bg-[#345073]"
+        navigation={navigation}
+      />
     </StyledView>
   );
 };
